app/db.py
from pymongo import MongoClient
import os
from dotenv import load_dotenv

load_dotenv()

mongo_uri = os.getenv("MONGO_URI")
client = MongoClient(mongo_uri, tls=True)

db = client["learningdb"]
goals_collection = db["goals"]
